Autorização configurações
Operações das configurações só podem ser realizadas por utilizadores com role de administrador.

+++ asset_management/app/Http/Controllers/CategoryController.php
@@ -34,6 +34,7 @@ class CategoryController extends Controller
 
     public function store(Request $request)
     {
+        $this->authorize('create-delete-users');
         $data = $request->validate([
             'name' => 'required|string|unique:categories,name',
         ]);
@@ -46,6 +47,7 @@ class CategoryController extends Controller
     //Delete Category
     public function destroy(Category $cat)
     {
+        $this->authorize('create-delete-users');
         $cat->delete();
 
         return response()->json(['message' => 'Category deleted successfully']);
@@ -54,6 +56,8 @@ class CategoryController extends Controller
     //Update Category
     public function update(Request $request, Category $category)
     {
+        $this->authorize('create-delete-users');
+
         $request->validate([
             'name' => 'required|string|max:255|unique:categories,name,' . $category->id,
         ]);
