+++ asset_management/app/Http/Controllers/EntityController.php
@@ -34,6 +34,7 @@ class EntityController extends Controller
 
     public function store(Request $request)
     {
+        $this->authorize('create-delete-users');
         $request->validate([
             'name' => 'required|string|max:255|unique:suppliers',
         ]);
@@ -47,6 +48,7 @@ class EntityController extends Controller
 
     public function destroy(Entity $ent)
     {
+        $this->authorize('create-delete-users');
         // Implement logic to delete the supplier
         // Don't forget to handle any related dependencies, if any
         $ent->delete();
@@ -57,6 +59,7 @@ class EntityController extends Controller
     //Update entity
     public function update(Request $request, Entity $ent)
     {
+        $this->authorize('create-delete-users');
         $request->validate([
             'name' => 'required|string|max:255|unique:categories,name,' . $ent->id,
         ]);
